Correct default for download
`landsat download LC80450332015023LGN00` should download the tar.gz file with all bands by default, but the default is being set to '432'. This does not do anything for scenes prior to 2015 on Google Storage, but for 2015 on AWS it only downloads those bands. This is not expected behaviour according to the docs, and it makes it hard to download the whole tar archive (you need to use `landsat download LC80450332015023LGN00 -b ""`)