Change default bands args to None to match docs
As reported in #191, Google Storage is not used to download Landsat 8 .tar.bz files even if they are available, as the list of bands is set to '' by default. Changing this to None fixes this.

I imagine the previous behaviour is not intentional, as it means that if no --bands argument is passed then the only files that are downloaded are the _MTL.txt and _BQA.tif files.